Fixed error in calculating simulationLocalityRate in report

+++ src/main/scala/tech/robins/SimulationReport.scala
@@ -21,5 +21,5 @@ case class SimulationReport(
 
   val totalLocalFetches: Int = taskExecutionReports.values.map(_.localResources).sum
   val totalRemoteFetches: Int = taskExecutionReports.values.map(_.remoteResources).sum
-  val simulationLocalityRate: Double = totalLocalFetches / (totalRemoteFetches + totalLocalFetches)
+  val simulationLocalityRate: Double = totalLocalFetches.toDouble / (totalRemoteFetches + totalLocalFetches)
 }
